complete authenticated routes

+++ src/routes/memories.ts
@@ -4,9 +4,12 @@ import { prisma } from "../lib/prisma"
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async request => await request.jwtVerify())
+
+    app.get('/memories', async request => {
 
         const memories = await prisma.memory.findMany({
+            where: { userId: request.user.sub },
             orderBy: { createAt: 'asc' }
         })
 
@@ -34,7 +37,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.post('/memories', async request => {
+    app.post('/memories', async (request, reply) => {
 
         const bodySchema = z.object({
             content: z.string(),
@@ -45,13 +48,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
         const memory = await prisma.memory.create({
-            data: { content, coverUrl, isPublic, userId: '86a7dffd-889e-46f1-91b1-e9eec19cc631' }
+            data: { content, coverUrl, isPublic, userId: request.user.sub }
         })
+
+        if (!memory.isPublic && memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         return memory
     })
 
-    app.put('/memories/:id', async request => {
+    app.put('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -67,7 +74,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: { id },
             data: { content, coverUrl, isPublic },
         })
@@ -75,13 +88,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async request => {
+    app.delete('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({ where: { id } })
     })
